refactor: extract createLangDropdown to Utils
