+++ src/Components/PageLayout/constants.PageLayout.ts
@@ -2,8 +2,6 @@ export const HEADER_CN = 'header';
 export const HEADER_TITLE_CN = `${HEADER_CN}--title`;
 export const HEADER_NAV_CN = `${HEADER_CN}--nav`;
 export const HEADER_BTN_CN = `${HEADER_CN}--btn`;
-export const HEADER_LANG_ITEM_CN = `${HEADER_CN}--lang_item`;
-export const HEADER_LANG_ICO_CN = `${HEADER_CN}--lang_ico`;
 
 export const MAIN_CN = 'main';
