+++ src/Components/PageLayout/index.PageLayout.ts
@@ -4,6 +4,7 @@ import { APP_NAME } from '../../Constants/index.Constants';
 import {
   createDropdown,
   createDropup,
+  createLangDropdown,
   createElement,
   createLink,
 } from '../../Utils/index.Utils';
@@ -14,15 +15,12 @@ import {
   MAIN_CN,
   HEADER_NAV_CN,
   HEADER_BTN_CN,
-  HEADER_LANG_ITEM_CN,
-  HEADER_LANG_ICO_CN,
   FOOTER_RS_LOGO_CN,
   FOOTER_YEAR_CN,
   FOOTER_BTN_CN,
 } from './constants.PageLayout';
 import { LOGIN, SETTINGS, STATISTICS } from '../../Constants/routes';
 import { PRIMARY_TEXT_CLASS, WRAPPER_CLASS } from '../../Constants/classNames';
-import languages from '../../langDictionaries/index.langDictionaries';
 import '../../assets/images/rs_logo.svg';
 
 const createUserDropdown = (observer: Observer, router: Router) => {
@@ -48,52 +46,6 @@ const createUserDropdown = (observer: Observer, router: Router) => {
   return userDropdown;
 };
 
-const createLangDropdown = (observer: Observer) => {
-  const { lang: currentLang, flag: currentFlag } = observer.getState().langData;
-
-  const currentIco = <HTMLImageElement>createElement('img', HEADER_LANG_ICO_CN);
-  currentIco.alt = currentLang;
-  currentIco.src = currentFlag;
-
-  const dropdownBtn = createElement(
-    'div',
-    [PRIMARY_TEXT_CLASS, HEADER_BTN_CN, HEADER_LANG_ITEM_CN],
-    null,
-    [currentIco],
-    currentLang
-  );
-
-  const langList = Object.values(languages);
-  const dropdownItems: Array<Element> = [];
-
-  langList.forEach((langDict) => {
-    const { lang, flag } = langDict;
-    if (lang === currentLang) return;
-
-    const ico = <HTMLImageElement>createElement('img', HEADER_LANG_ICO_CN);
-    ico.alt = lang;
-    ico.src = flag;
-
-    const dropdownItem = createElement(
-      'div',
-      [PRIMARY_TEXT_CLASS, HEADER_LANG_ITEM_CN],
-      null,
-      [ico],
-      lang
-    );
-
-    dropdownItem.addEventListener('click', () => {
-      observer.actions.setLang(langDict);
-    });
-
-    dropdownItems.push(dropdownItem);
-  });
-
-  const dropdown = createDropdown(dropdownBtn, dropdownItems);
-
-  return dropdown;
-};
-
 const createHeader = (observer: Observer, router: Router) => {
   const wrapper = createElement('div', WRAPPER_CLASS);
 
@@ -108,7 +60,7 @@ const createHeader = (observer: Observer, router: Router) => {
 
   const menuContainer = createElement('nav', HEADER_NAV_CN);
 
-  const langDropdown = createLangDropdown(observer);
+  const langDropdown = createLangDropdown(observer, HEADER_BTN_CN);
   menuContainer.append(langDropdown);
 
   const userDropdown = createUserDropdown(observer, router);
