+++ src/styles/main.scss
@@ -141,6 +141,19 @@ body {
     margin-top: -$border_thickness;
   }
 
+  &--lang_item {
+    padding: 0 1em;
+    box-sizing: border-box;
+    display: flex;
+    justify-content: space-between;
+    gap: 0.5em;
+  }
+
+  &--lang_ico {
+    width: 20px;
+    height: 20px;
+  }
+
   &:hover {
     .dropdown--btn {
       cursor: pointer;
